requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
ruff>=0.4.0
